pyproject.toml
[project]
name = "aiagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
    "uvicorn[standard]>=0.35.0",
]

[tool.uv.workspace]
members = [
    "aiagent",
]
